requirements.txt
# Talent Reconnect Agent - Core Dependencies
# Install with: pip install -r requirements.txt --pre

# Environment & Utilities
python-dotenv>=1.0.0

# Microsoft Agent Framework (Preview)
agent-framework>=1.0.0b251114
agent-framework-azure-ai>=1.0.0b251114
agent-framework-azure-ai-search>=1.0.0b251114

# Azure Authentication & LLM Client
azure-identity>=1.19.0
openai>=1.50.0

# Azure AI Search (for resume sourcing with hybrid search)
azure-search-documents>=11.6.0
aiohttp>=3.9.0

# Web Framework (API mode)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Web UI (Chainlit - optional)
chainlit>=1.0.0
plotly>=5.0.0
# Data Generation (for synthetic resumes)
faker>=18.0.0
pandas>=2.0.0